# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/symbols": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Adds a new trading symbol to the system (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Symbols"
                ],
                "summary": "Create a new symbol",
                "parameters": [
                    {
                        "description": "Symbol data",
                        "name": "symbol",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Symbol"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Symbol created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to create symbol",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/symbols/{id}": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Updates the details of an existing trading symbol (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Symbols"
                ],
                "summary": "Update a symbol",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Symbol ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated symbol data",
                        "name": "symbol",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Symbol"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Symbol updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to update symbol",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Removes a trading symbol from the system (admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Symbols"
                ],
                "summary": "Delete a symbol",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Symbol ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Symbol deleted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to delete symbol",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/symbols": {
            "get": {
                "description": "Retrieves a list of all trading symbols",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Symbols"
                ],
                "summary": "Get all symbols",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Symbol"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve symbols",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/symbols/{id}": {
            "get": {
                "description": "Retrieves details of a trading symbol by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Symbols"
                ],
                "summary": "Get symbol by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Symbol ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Symbol"
                        }
                    },
                    "400": {
                        "description": "Invalid symbol ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Symbol not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/signup": {
            "post": {
                "description": "Registers a new user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User account data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserAccount"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to create user",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Retrieves user account details by user ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserAccount"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Symbol": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "commission": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "denied_accounts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "display_name": {
                    "type": "string"
                },
                "is_trading_open": {
                    "type": "boolean"
                },
                "leverage": {
                    "type": "integer"
                },
                "max_lot": {
                    "type": "number"
                },
                "min_lot": {
                    "type": "number"
                },
                "spread": {
                    "type": "number"
                },
                "symbol_name": {
                    "type": "string"
                },
                "trading_hours": {
                    "$ref": "#/definitions/models.TradingHours"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.TradingHours": {
            "type": "object",
            "properties": {
                "close_time": {
                    "type": "string"
                },
                "open_time": {
                    "type": "string"
                },
                "unlimited": {
                    "type": "boolean"
                }
            }
        },
        "models.UserAccount": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "account_name": {
                    "type": "string"
                },
                "account_type": {
                    "type": "string"
                },
                "balance": {
                    "type": "number"
                },
                "bonus": {
                    "type": "number"
                },
                "card_number": {
                    "type": "string"
                },
                "citizenship": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "leverage": {
                    "type": "integer"
                },
                "national_id": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "registration_date": {
                    "type": "string"
                },
                "telegram_id": {
                    "type": "string"
                },
                "trade_type": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "wallet_address": {
                    "type": "string"
                }
            }
        }
    }
}